+++ src/main/java/ui/message/InputMessage.java
@@ -1,0 +1,17 @@
+package ui.message;
+
+public enum InputMessage {
+    INPUT_PURCHASE_PRICE("구입금액을 입력해 주세요."),
+    INPUT_LAST_WEEK_WINNING_NUMBERS("\n지난 주 당첨 번호를 입력해 주세요."),
+    INPUT_BONUS_BALL("보너스 볼을 입력해 주세요.");
+
+    private String message;
+
+    InputMessage(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
