+++ src/main/java/ui/receiver/InputValidator.java
@@ -5,7 +5,7 @@ import java.util.Set;
 
 import domain.lotteryService.BonusNumber;
 import domain.lotteryService.WinningNumbers;
-import ui.Message;
+import ui.message.ExceptionMessage;
 
 public class InputValidator {
     private static final int MINIMUM_PURCHASE_PRICE = 1000;
@@ -15,7 +15,7 @@ public class InputValidator {
 
     public static void validatePurchasePrice(int purchasePrice) {
         if (purchasePrice < MINIMUM_PURCHASE_PRICE) {
-            throw new IllegalArgumentException(Message.ExceptionMessage.MUST_BUY_MORE_THAN_ONE_TICKET.getMessage());
+            throw new IllegalArgumentException(ExceptionMessage.MUST_BUY_MORE_THAN_ONE_TICKET.getMessage());
         }
     }
 
@@ -25,7 +25,7 @@ public class InputValidator {
 
     public static void validateSizeOfWinningNumbers(WinningNumbers winningNumbers) {
         if (winningNumbers.size() != 6) {
-            throw new IllegalArgumentException(Message.ExceptionMessage.MUST_INPUT_SIX_WINNING_NUMBERS.getMessage());
+            throw new IllegalArgumentException(ExceptionMessage.MUST_INPUT_SIX_WINNING_NUMBERS.getMessage());
         }
     }
 
@@ -34,28 +34,28 @@ public class InputValidator {
             .stream()
             .anyMatch(
                 winningNumber -> winningNumber < MINIMUM_WINNING_NUMBER || winningNumber > MAXIMUM_WINNING_NUMBER)) {
-            throw new IllegalArgumentException(Message.ExceptionMessage.MUST_INPUT_NUMBERS_IN_VALID_RANGE.getMessage());
+            throw new IllegalArgumentException(ExceptionMessage.MUST_INPUT_NUMBERS_IN_VALID_RANGE.getMessage());
         }
     }
 
     public static void validateRangeOfBonusNumber(BonusNumber bonusNumber) {
         if (bonusNumber.getBonusNumber() < MINIMUM_WINNING_NUMBER
             || bonusNumber.getBonusNumber() > MAXIMUM_WINNING_NUMBER) {
-            throw new IllegalArgumentException(Message.ExceptionMessage.MUST_INPUT_NUMBERS_IN_VALID_RANGE.getMessage());
+            throw new IllegalArgumentException(ExceptionMessage.MUST_INPUT_NUMBERS_IN_VALID_RANGE.getMessage());
         }
     }
 
     public static void validateDuplicateWinningNumbers(WinningNumbers winningNumbers) {
         Set<Integer> numbers = new HashSet<>(winningNumbers.getWinningNumbers());
         if (numbers.size() != SIZE_OF_WINNING_NUMBERS) {
-            throw new IllegalArgumentException(Message.ExceptionMessage.DUPLICATE_NUMBERS.getMessage());
+            throw new IllegalArgumentException(ExceptionMessage.DUPLICATE_NUMBERS.getMessage());
         }
     }
 
     public static void validateBonusNumber(WinningNumbers winningNumbers, BonusNumber bonusNumber) {
         if (winningNumbers.contains(bonusNumber.getBonusNumber())) {
             throw new IllegalArgumentException(
-                Message.ExceptionMessage.BONUS_NUMBER_CANNOT_EQUAL_WINNING_NUMBERS.getMessage());
+                ExceptionMessage.BONUS_NUMBER_CANNOT_EQUAL_WINNING_NUMBERS.getMessage());
         }
     }
 }
